Make code backward compatible
Previous commit was enforcing local activities to be registered into the
worker to be ran. However, current behavior is that code like

ExecuteLocalActivity(ctx, func() error {
    return nil
})

is actually valid meaning local activity registration is optional when
you pass down the activity function directly.

This commits bring backs the old ExecuteLocalActivity behavior when the
activity passed is a func and tries to load the activity func from the
registry when we pass a string.